requirements-dev.txt
flake8==3.7.8
django-nose==1.4.6
coverage==4.5.4
isort==4.3.21
black==19.10b0